// src/generated/graphql.tsx
export type GetWorkoutsQuery = { __typename?: 'query_root', workouts: Array<{ __typename?: 'workouts', id: number, name?: string | null, created_at: any, updated_at: any, sets: Array<{ __typename?: 'sets', completed_at?: any | null, created_at: any, exercise_id: number, id: number, repetitions: number, type: string, updated_at: any, weight: any, exercise: { __typename?: 'exercises', name: string } }> }> };